requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==20.7
websockets>=12.0
